Implementing shorter min-max Implementing a min-max algorithm that exploits the state.value() in a better way. state.value() calculates the loss of the current turn's player.
min_max_two will then minimaze the loss if its the starting player's turn,
and maximize it if it's the opposing player turn. We keep track of which player
started by using the opposing_player argument.

This is necessary, because state.current_player is true if it's
the red player's turn, and false if it's the blue player's turn.
If we want to give to the blue player's the minmax Strategy, there would
be no way of knowing whether we should minimaze or maximize.